refactor(gateway): isolate accepted workspace lock
Keep the byte-identical remote lock fragment with the accepted-workspace transaction instead of the manifest implementation.
